Add benchamrk for Go's copy method
The old version that stopped and started the timer was doing something
sketchy and running the benchmarks with just one element basically
panicked, thus I've removed those calls and added a benchmark for Go's
copy method, so users can substract it from the results.

Not perfect, I know, but it's something.